move and rename globals and contract class
